\name{bam2metaseqr}
\alias{bam2metaseqr}
\title{metaseqr pipeline starting from raw BAM or BED files}
\usage{
  bam2metaseqr(...)
}
\arguments{
  \item{...}{not yet implemented}
}
\value{
  The same output as the main \code{metaseqr} pipeline.
}
\description{
  This function is a wrapper for the \code{metaseqr}
  pipeline, starting from BAM or BED files instead of a
  read counts file.
}
\examples{
# Not yet implenented
}
\author{
  Panagiotis Moulos
}

