\name{check.contrast.format}
\alias{check.contrast.format}
\title{Contrast validator}
\usage{
  check.contrast.format(cnt, sample.list)
}
\arguments{
  \item{cnt}{contrasts vector.}

  \item{sample.list}{the input sample list.}
}
\description{
  Checks if the contrast vector follows the specified
  format. Internal use only.
}
\author{
  Panagiotis Moulos
}

