\name{check.file.args}
\alias{check.file.args}
\title{File argument validator}
\usage{
  check.file.args(arg.name, arg.value)
}
\arguments{
  \item{arg.name}{argument name to display in a possible
  error.}

  \item{arg.value}{the filename to check.}
}
\description{
  Checks if a file exists for specific arguments requiring
  a file input. Internal use only.
}
\author{
  Panagiotis Moulos
}

