\name{check.libsize}
\alias{check.libsize}
\title{Library size validator}
\usage{
  check.libsize(libsize.list, sample.list)
}
\arguments{
  \item{libsize.list}{the samples-names library size list.}

  \item{sample.list}{the input sample list.}
}
\description{
  Checks the names of the supplied library sizes. Internal
  use only.
}
\author{
  Panagiotis Moulos
}

