\name{check.packages}
\alias{check.packages}
\title{Required packages validator}
\usage{
  check.packages(a, m, s, p)
}
\arguments{
  \item{a}{multiple testing correction method.}
  
  \item{m}{meta-analysis method.}

  \item{s}{data transformation method.}
  
  \item{p}{QC plot types.}
}
\description{
  Checks if all the any required packages, not attached 
  during installation or loading, are present according 
  to metaseqR input options. Internal use only.
}
\author{
  Panagiotis Moulos
}

