\name{check.parallel}
\alias{check.parallel}
\title{Parallel run validator}
\usage{
  check.parallel(rc)
}
\arguments{
  \item{rc}{fraction of available cores to use.}
}
\description{
  Checks existence of multiple cores and loads multicore
  package. Internal use only.
}
\author{
  Panagiotis Moulos
}

