\name{filter.high}
\alias{filter.high}
\title{Filtering helper}
\usage{
  filter.high(x, f)
}
\arguments{
  \item{x}{a data numeric matrix.}

  \item{f}{a threshold.}
}
\description{
  High score filtering function. Internal use.
}
\examples{
\dontrun{
filter.high(data,median(data))
}
}
\author{
  Panagiotis Moulos
}

