\name{filter.low}
\alias{filter.low}
\title{Filtering helper}
\usage{
  filter.low(x, f)
}
\arguments{
  \item{x}{a data numeric matrix.}

  \item{f}{a threshold.}
}
\description{
  Low score filtering function. Internal use.
}
\examples{
\dontrun{
filter.low(data,median(data))
}
}
\author{
  Panagiotis Moulos
}

