\docType{data}
\name{libsize.list.hg18}
\alias{libsize.list.hg18}
\title{Human RNA-Seq data with two conditions, four samples}
\format{a named \code{list} with library sizes.}
\source{
  BSRC Alexander Fleming (http://www.fleming.gr)
}
\description{
  The library size list for \code{hg18.exon.counts}. See
  the data set description.
}
\author{
  Panagiotis Moulos
}
\keyword{datasets}

