\docType{data}
\name{libsize.list.mm9}
\alias{libsize.list.mm9}
\title{Mouse RNA-Seq data with three conditions, six samples}
\format{a named \code{list} with library sizes.}
\source{
  BSRC Alexander Fleming (http://www.fleming.gr)
}
\description{
  The library size list for \code{mm9.gene.counts}. See the
  data set description.
}
\author{
  Panagiotis Moulos
}
\keyword{datasets}

