\docType{package}
\name{metaseqR-package}
\alias{metaseqR}
\alias{metaseqR-package}
\title{The metaseqR Package}
\description{
  An R package for the analysis and result reporting
  of RNA-Seq gene expression data, using multiple 
  statistical algorithms.
}
\details{
  \tabular{ll}{ Package: \tab metaseqr\cr Type: \tab
  Package\cr Version: \tab 0.9.1\cr Date: \tab
  2013-12-06\cr Depends: \tab R (>= 2.13.0), Biobase,
  BiocGenerics, rjson, biomaRt, utils, knitr, EDASeq,
  DESeq,\cr edgeR, limma, NOISeq, baySeq, NBPSeq, MADAM,
  survcomp, brew, gplots,\cr corrplot, GenomicRanges,
  Rsamtools, rtracklayer, Repitools, qvalue, vsn,
  VennDiagram, log4r\cr Encoding: \tab UTF-8\cr License: \tab GPL
  (>= 3)\cr LazyLoad: \tab yes\cr URL: \tab
  http://www.fleming.gr\cr }

  Provides an interface to several normalization and
  statistical testing packages for RNA-Seq gene expression
  data. Additionally, it creates several diagnostic plots,
  performs meta-analysis by combinining the results of
  several statistical tests and reports the results in an
  interactive way.
}
\author{
  Panagiotis Moulos \email{moulos@fleming.gr}
}
\keyword{package}

