\docType{data}
\name{mm9.gene.counts}
\alias{mm9.gene.counts}
\title{mouse RNA-Seq data with three conditions, six samples}
\format{a \code{data.frame} with gene read counts and some embedded annotation, one row per gene.}
\source{
  BSRC Alexander Fleming (http://www.fleming.gr)
}
\description{
  This data set contains RNA-Seq gene read counts for 3
  chromosomes. The data are from an experiment studying
  gene expression at several developmental stages of mouse
  liver. It has three conditions-developmental stages
  (e15.5, P0.5, P60) and six samples (e15.5_1, e15.5_2,
  P0.5_1, P0.5_2, P60_1, P60_2). It also contains a
  predefined \code{sample.list} and \code{libsize.list}
  named \code{sample.list.mm9} and \code{libsize.list.mm9}.
}
\author{
  Panagiotis Moulos
}
\keyword{datasets}

