\docType{data}
\name{sample.list.hg18}
\alias{sample.list.hg18}
\title{Human RNA-Seq data with two conditions, four samples}
\format{a named \code{list} with condition and sample names.}
\source{
  BSRC Alexander Fleming (http://www.fleming.gr)
}
\description{
  The sample list for \code{hg18.exon.counts}. See the data
  set description.
}
\author{
  Panagiotis Moulos
}
\keyword{datasets}

