\name{check.packages}
\alias{check.packages}
\title{Required packages validator}
\usage{
  check.packages(n, d, a, m, s, p, r)
}
\arguments{
  \item{n}{normalization method}

  \item{d}{statistics method}

  \item{a}{multiple testing correction method}

  \item{m}{meta-analysis method}

  \item{s}{data transformation method}

  \item{p}{qc plot types}

  \item{r}{reporting option}
}
\description{
  Checks if all the required packages are present according
  to metaseqr input options. Internal use only.
}
\author{
  Panagiotis Moulos
}

